+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Changelog for NeoFS Node
 ### Changed
 - SN retries notary requests if `insufficient amount of gas` error appears (#3739)
 - Speed up metabase resync by using batch operations (#3804)
+- Session key storage access session keys by account only (#3817)
 
 ### Removed